blah

+++ tasks/clean.rake
@@ -20,6 +20,8 @@ task :clean, [:del] do |t,args|
     cputs "Removing packaged VMs"
     FileUtils.rm_rf(Dir.glob(CACHEDIR+"/vms/*.zip*"))
   end
-  cputs "Archiving post.log"
-  FileUtils.mv CACHEDIR+"/post.log", CACHEDIR+"/post.log_lastrun", :force => true
+  if File.exist?(CACHEDIR+"/post.log")
+    cputs "Archiving post.log"
+    FileUtils.mv CACHEDIR+"/post.log", CACHEDIR+"/post.log_lastrun", :force => true
+  end
 end
